requirements.txt
asgiref==3.4.1
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
fastapi==0.68.0
h11==0.12.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
llama-cpp-python
pyllamacpp